+++ Admin/register/admin.register.js
@@ -130,21 +130,94 @@ function closeForm(onfinish) {
 }
 
 
+class FormSet {
+  constructor(state = formsID[0], data = {}, completed = false) {
+    this.state = state;
+    this.data = data;
+    if (!this.data.isAdmin) {
+      this.set('isAdmin', true)
+    }
+    if (!this.data.isPermanent) {
+      this.set('isPermanent', true)
+    }
+    this.completed = completed;
+  }
+
+  set(key, name) {
+    this.data[key] = name;
+  }
+
+  get(key) {
+    return this.data[key];
+  }
+
+  index() {
+    if (formsID) {
+      return formsID.indexOf(this.state)
+    }
+  }
+}
+
+
+
 document.getElementById('nextBtn').onclick = function() {
-  const key = 'ADMN_' + Math.floor(Math.random() * 9999) + '_AG' + (['ABV', 'AKC', 'ZSO', 'KML', 'OPT'][Math.floor(Math.random() * 5)]) + '_AWq67/' + (['Uwes5', 'Ksn74', 'amFw8n', 'Wesy6', '7she73'][Math.floor(Math.random() * 5)]);
-  
-  caches.open('admin').then(function(cache) {
-    cache.put('admin-access-key', new Response(JSON.stringify({ key }), {
-      headers: { 'Content-type': 'application/json' }
-    })).catch((err) => {
-      alert(err)
-    }).then(() => {
-      cache.keys('admin-access-key').then(function(t) {
-        localStorage.setItem('adminUrl', t[0].url)
-      })
+  var form_set = new FormSet('default');
+  var name = document.getElementById('fullname'),
+    adminKeyInp = document.getElementById('key'),
+    password = document.getElementById('password');
+
+  if (!name.value) {
+    changeLog('name input is blank.')
+  } else if (!adminKeyInp.value) {
+    changeLog('admin key is wrong.')
+  } else if (!password.value) {
+    changeLog('personal password input is blank');
+  } else {
+    form_set.set('fullname', name.value);
+    form_set.set('adminKey', adminKeyInp.value);
+    var encrypted = CryptoJS.AES.encrypt(password.value, config.ENC_KEY).toString();
+    form_set.set('password', encrypted);
+
+    if (typeof spinner != 'undefined') {
+      spinner.showPreloader('Verifying...');
+    }
+    bushido.useQuery('adminKey', [
+      ['key', '==', adminKeyInp.value]
+      ]).then(function(snapshot) {
+      var data = bushido.toData(snapshot);
+      if (data.length == 0) {
+        alert("doesn't match admin key")
+      } else {
+        var obj = data[0].data()
+        if (obj.key == adminKeyInp.value) {
+          const key = 'ADMN_' + Math.floor(Math.random() * 9999) + '_AG' + (['ABV', 'AKC', 'ZSO', 'KML', 'OPT'][Math.floor(Math.random() * 5)]) + '_AWq67/' + (['Uwes5', 'Ksn74', 'amFw8n', 'Wesy6', '7she73'][Math.floor(Math.random() * 5)]);
+
+          if (typeof spinner != 'undefined') {
+            spinner.changeText('Verified. Connecting to server...');
+          }
+          bushido.set(('accounts/ADMN_' + Math.floor(Math.random() * 999999)), form_set.data).then(() => {
+            caches.open('admin').then(function(cache) {
+              cache.put('admin-access-key', new Response(JSON.stringify({ key }), {
+                headers: { 'Content-type': 'application/json' }
+              })).catch((err) => {
+                alert(err)
+              }).then(() => {
+                cache.keys('admin-access-key').then(function(t) {
+                  localStorage.setItem('adminUrl', t[0].url)
+                  window.location.href = '../attendance';
+                })
+              })
+            }).catch((err) => {
+              alert(err)
+            });
+            localStorage.setItem('adminKey', key)
+          }).catch((err) => {
+            alert(err)
+          });
+        } else {
+          alert("doesn't match admin key")
+        }
+      }
     })
-  }).catch((err) => {
-    alert(err)
-  });
-  localStorage.setItem('adminKey', key)
+  }
 }